# Lobe pruning shader optimization