… lib/mayaUsd/render/MaterialXGenOgsXml/ShaderGenUtil.h
… lib/mayaUsd/render/vp2RenderDelegate/material.cpp
… lib/mayaUsdAPI/render.cpp
… test/lib/mayaUsd/utils/test_ShaderGenUtils.cpp